requirements.txt
pipe~=2.0
numpy~=1.26.2
Pillow~=10.1.0
tifffile~=2023.9.26
